Use PROJECT_VERSION for VulkanHeaders
Uses PROJECT_VERSION in place of CMAKE_PROJECT_VERSION for the desired
version of the VulkanHeaders package. The latter is not useful if the
repository is used as a submodule, as it will take on the value of the
top-level project's version. The former fixes this issue.

Also updates the project version to 1.4.328.